Update README.md
This is a repository for Python programs related to the number format. It contains programs that can be used to manipulate numbers, perform basic calculations, format numbers for different purposes, and more. The repository offers a wide variety of programs to meet the needs of any user and is designed to be easy to use and understand. For those looking to improve their understanding of number format, this repository is an ideal resource.